// src/app/Pages/Models/pincode.ts
export class pincode {
    ID: number;
    OFFICE_NAME: string = '';
    PINCODE: string;
    DIVISION_NAME: any;
    CIRCLE_NAME: any;
    TALUKA: any;
    DISTRICT: any;
    PINCODE_NUMBER: any
    STATE: any;
    STATE_ID: any;
    COUNTRY_ID: any;
    SUB_OFFICE: any;
    HEAD_OFFICE: any;
    LONGITUDE: any;
    LATTITUDE: any;
    IS_ACTIVE: Boolean = true;
    CLIENT_ID: any;
    COUNTRY_NAME: any;
    STATE_NAME: any;
    DISTRICT_NAME: any;
    PINCODE_FOR: any;
}